+++ tests/covalent_dispatcher_tests/_cli/groups/db_test.py
@@ -18,7 +18,7 @@ from unittest.mock import Mock
 
 from click.testing import CliRunner
 
-from covalent_dispatcher._cli.groups.db import MIGRATION_WARNING_MSG, alembic, migrate
+from covalent_dispatcher._cli.groups.db_group import MIGRATION_WARNING_MSG, alembic, migrate
 from covalent_dispatcher._db.datastore import DataStore
 
 
@@ -46,7 +46,9 @@ def test_alembic_command_args(mocker):
     MOCK_ALEMBIC_ARGS_INVALID = "some alembic command --with-flags -m 'comment'"
     ALEMBIC_ERROR_STDERR = b"alembic: error: invalid..."
     ALEMBIC_ERROR_STDOUT = b"b60c5 (head)"
-    popen_mock = mocker.patch.object(sys.modules["covalent_dispatcher._cli.groups.db"], "Popen")
+    popen_mock = mocker.patch.object(
+        sys.modules["covalent_dispatcher._cli.groups.db_group"], "Popen"
+    )
     # test valid alembic args
     popen_mock().communicate.return_value = (ALEMBIC_ERROR_STDOUT, b"")
     res = runner.invoke(alembic, MOCK_ALEMBIC_ARGS_VALID, catch_exceptions=False)
@@ -61,7 +63,9 @@ def test_alembic_not_installed_error(mocker):
     runner = CliRunner()
     MOCK_ALEMBIC_ARGS = "current"
     EXCEPTION_MESSAGE = "[Errno 2] No such file or directory: 'alembic'"
-    popen_mock = mocker.patch.object(sys.modules["covalent_dispatcher._cli.groups.db"], "Popen")
+    popen_mock = mocker.patch.object(
+        sys.modules["covalent_dispatcher._cli.groups.db_group"], "Popen"
+    )
     popen_mock.side_effect = FileNotFoundError(EXCEPTION_MESSAGE)
     res = runner.invoke(alembic, MOCK_ALEMBIC_ARGS, catch_exceptions=False)
     assert EXCEPTION_MESSAGE in res.output