+++ CHANGELOG.md
@@ -7,13 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [UNRELEASED]
 
+### Changed
+
+- Terraform output to use scrolling buffer.
+- Terraform output handling to show errors.
+
 ## [0.231.0-rc.0] - 2023-11-28
 
 ### Authors
 
 - Ara Ghukasyan <38226926+araghukas@users.noreply.github.com>
 - Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
-
 
 ### Added